Fix string input in plotAll, create run script
plotAll now takes SampleOrderString as input and tokenizes it properly
to get the sample names. I fixed the usage of strtok and fixed some
memory issues that were occurring. It should now be easier to
incorporate this into a condor job.

I also created run_plotAll.C as a run script for plotAll. The same thing
can be accomplished with a shell script as well - here it doesn't matter
since it exists only for testing.